requirements.txt
django==4.2.3
psycopg2-binary==2.9.6
pillow==10.0.0
django-cors-headers==4.2.0
python-decouple==3.8
django-phonenumber-field[phonenumbers]==7.1.0
django-filter==23.2
djangorestframework==3.14.0
drf_spectacular==0.26.4
djangorestframework-simplejwt==5.2.2
django-smtp-ssl==1.0
django-mptt==0.14.0
django-rest-passwordreset==1.3.0
celery==5.2.7
channels-redis==4.1.0
channels[daphne]==4.0.0

